add a separate viewState for submitting addons
